initialize invoice controls

+++ src/main/java/com/jh/hotelbookingmanagement/entity/Promotion.java
@@ -18,7 +18,7 @@ public class Promotion {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "Promotion_ID")
-    Integer promotionId;
+    Long promotionId;
 
     @Column(name = "Promotion_Name")
     String promotionName;
